requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
blosc2==4.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.5
cycler==0.12.1
einops==0.8.2
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.46
idna==3.11
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
knn-cuda @ https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl
lightning-utilities==0.15.3
markupsafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
ndindex==1.10.1
networkx==3.4.2
ninja==1.13.0
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.9.2
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pyg-lib==0.5.0+pt28cu128
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.53.0
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tables==3.10.1
threadpoolctl==3.6.0
torch==2.8.0+cu128
torch-cluster==1.6.3+pt28cu128
torch-geometric==2.7.0
torch-scatter==2.1.2+pt28cu128
torch-sparse==0.6.18+pt28cu128
torch-spatiotemporal==0.9.5
torch-spline-conv==1.2.2+pt28cu128
torchaudio==2.8.0+cu128
torchmetrics==1.8.2
torchvision==0.23.0+cu128
tqdm==4.67.3
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
xxhash==3.6.0
yarl==1.22.0
